MovieHubNovember/MovieHubNovember/urls.py
"""MovieHubNovember URL Configuration

The `urlpatterns` list routes URLs to views. For more information please see:
    https://docs.djangoproject.com/en/3.2/topics/http/urls/
Examples:
Function views
    1. Add an import:  from my_app import views
    2. Add a URL to urlpatterns:  path('', views.home, name='home')
Class-based views
    1. Add an import:  from other_app.views import Home
    2. Add a URL to urlpatterns:  path('', Home.as_view(), name='home')
Including another URLconf
    1. Import the include() function: from django.urls import include, path
    2. Add a URL to urlpatterns:  path('blog/', include('blog.urls'))
"""
from django.contrib import admin
from django.urls import path
from rest_framework.routers import DefaultRouter
from movieapi import views as api_view
from rest_framework.authtoken.views import ObtainAuthToken
from rest_framework_simplejwt.views import TokenObtainPairView,TokenRefreshView

from rest_framework import permissions
from drf_yasg.views import get_schema_view
from drf_yasg import openapi

from django.conf import settings
from django.conf.urls.static import static

schema_view = get_schema_view(
   openapi.Info(
      title="Snippets API",
      default_version='v1',
      description="Test description",
      terms_of_service="https://www.google.com/policies/terms/",
      contact=openapi.Contact(email="contact@snippets.local"),
      license=openapi.License(name="BSD License"),
   ),
   public=True,
   permission_classes=[permissions.AllowAny],
)





router=DefaultRouter()
router.register("api/register",api_view.UsersView,basename="users")
router.register("api/movies",api_view.MoviesView,basename="movies")
router.register("api/reviews",api_view.ReviewsView,basename="reviews")



urlpatterns = [
    path('admin/', admin.site.urls),
    path("api/token/",ObtainAuthToken.as_view()),
    path("api/v1/token/",TokenObtainPairView.as_view()),
    path("api/v1/token/refresh/",TokenRefreshView.as_view()),
    path('swagger/', schema_view.with_ui('swagger', cache_timeout=0), name='schema-swagger-ui'),
]+router.urls+ static(settings.MEDIA_URL, document_root=settings.MEDIA_ROOT)
